% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{chapter_redoc}
\alias{chapter_redoc}
\title{Creates an R Markdown Word Thesis chapter that can be read back into R markdown}
\usage{
chapter_redoc(...)
}
\arguments{
\item{...}{other arguments to  bookdown::word_document2}
}
\value{
A Word Document somewhat based on the guidelines, but needs manual editing.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the chapter using the redoc package.
Very very preliminary.
}
\examples{
\dontrun{
 output: crimsondown::chapter_redoc
}
}
\seealso{
\url{https://noamross.github.io/redoc/} for more about the redoc package
}
