% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{chapter_word}
\alias{chapter_word}
\title{Creates an R Markdown Word Thesis chapter with more formatting options}
\usage{
chapter_word(...)
}
\arguments{
\item{...}{other arguments to bookdown::word_document2}
}
\value{
A Word Document of a single chapter
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the thesis using the officedown package.
}
\examples{
\dontrun{
 output: crimsondown::chapter_word
}
}
\seealso{
\url{https://github.com/davidgohel/officedown} for more about the officedown package
}
