% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimsondown.R
\docType{package}
\name{crimsondown}
\alias{crimsondown}
\title{crimsondown: A package for writing theses that conform to Harvard University
Graduate SChool of Arts and Sciences dissertation formatting rules}
\description{
A package for writing theses that conform to Harvard University
Graduate SChool of Arts and Sciences dissertation formatting rules
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\section{chapter_pdf}{

Creates an R Markdown thesis chapter as a PDF
}

\section{chapter_word}{

Creates an R Markdown thesis chapter as a Microsoft Word document


Creates an R Markdown thesis chapter as a Microsoft Word "redoc" document
}

